updated vehicle population chart
