+++ Visualization/Web/Web.Api/WebUI/src/app/components/dashboard/charts/vehicle-population/vehicle-population.component.css
@@ -1,0 +1,24 @@
+
+.chart-card {
+    width: 800px;
+}
+
+.chart-header {
+    margin-bottom: -20px;
+}
+
+.chart-title {
+    margin: 0 -10px;
+}
+
+hr {
+    margin-left: -16px;
+    width: calc(100% + 32px);
+}
+
+.chart {
+    margin-top: 10px;
+    width: calc(100% + 16px);
+    height: 30%;
+    margin: 0 -8px 8px -8px;
+}
