+++ Visualization/Web/Web.Api/WebUI/src/app/components/dashboard/charts/vehicle-population/vehicle-population.component.html
@@ -1,4 +1,4 @@
-<mat-card class="chart-card" *ngIf="vehicleChart.chartLabels.length > 0">
+<mat-card class="chart-card">
         <mat-card-header class="chart-header">
                 <div mat-card-avatar>
                     <i class="material-icons icon">insert_chart_outlined</i>
