+++ Visualization/Web/Web.Api/WebUI/src/app/components/dashboard/dashboard.component.css
@@ -28,25 +28,8 @@
     font-size: 30px;
 }
 
-:host >>> .chart-card {
-    width: 500px;
-}
-
-:host >>> .chart-header {
-    margin-bottom: -20px;
-}
-
-:host >>> .chart-title {
-    margin: 0 -10px;
-}
-
-:host >>> hr {
-    margin-left: -16px;
-    width: calc(100% + 32px);
-}
-
-:host >>> .chart {
-    margin-top: 10px;
-    width: calc(100% + 16px);
-    margin: 0 -8px 8px -8px;
+.stats-container {
+    display: flex;
+    justify-content: space-between;
+    flex-wrap: wrap;
 }
