+++ Visualization/Web/Web.Api/WebUI/src/app/components/dashboard/dashboard.component.html
@@ -1,6 +1,8 @@
 <div class="container-flex margin-top-lg">
     <mat-accordion multi="true">
-        <mat-expansion-panel [expanded]="true">
+        <mat-expansion-panel>
+        <!-- [expanded]="true" -->
+        
                 <mat-expansion-panel-header class="header">
                     Podgląd mapy
                 </mat-expansion-panel-header>
@@ -9,14 +11,16 @@
                 </div>
             </mat-expansion-panel>
             
-            <mat-expansion-panel>
+            <mat-expansion-panel [expanded]="true">
                 <mat-expansion-panel-header class="header">
                     Statystyki
                 </mat-expansion-panel-header>
-                <ng-template matExpansionPanelContent *ngIf="connected; else loading">
-                    <app-statistics></app-statistics>
-                    <app-vehicle-population></app-vehicle-population>
-                </ng-template>
+                <div matExpansionPanelContent *ngIf="connected; else loading">
+                    <div class="stats-container">
+                        <app-statistics></app-statistics>
+                        <app-vehicle-population></app-vehicle-population>
+                    </div>
+                </div>
                 
             </mat-expansion-panel>
         </mat-accordion>
