+++ Visualization/Web/Web.Api/WebUI/src/app/components/dashboard/stats/statistics/statistics.component.css
@@ -1,0 +1,3 @@
+table {
+    min-width: 500px;
+}
