+++ Visualization/Web/Web.Api/WebUI/src/app/components/unity-connector/unity-connector.component.css
@@ -61,7 +61,7 @@ svg:hover {
     margin: 5px;
 }
 
-@media screen and (max-width: 1110px) {
+@media screen and (max-width: 1145px) {
     .simulator-title {
         display: none;
     }
